src/main/java/com/storemanagementspring/controllers/ProductController.java
package com.storemanagementspring.controllers;

import com.storemanagementspring.dto.ProductDTO;
import com.storemanagementspring.models.Product;
import com.storemanagementspring.service.ProductService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("store-management/products")
public class ProductController {

    private ProductService productService;

    public ProductController(ProductService productService) {
        this.productService = productService;
    }

    @PostMapping()
    public ResponseEntity<ProductDTO> addProduct(ProductDTO productDTO){
        this.productService.addProduct(productDTO);
        return new ResponseEntity<>(productDTO, HttpStatus.CREATED);
    }

    @GetMapping()
    public ResponseEntity<List<Product>> getProducts(){
        return new ResponseEntity<>(this.productService.getProducts(), HttpStatus.OK);
    }

    @GetMapping("/{id}")
    public ResponseEntity<Product> getProductById(@PathVariable Long id){
        return new ResponseEntity<>(this.productService.getProductById(id), HttpStatus.OK);
    }
}
